Add generic chunked storage to strings and numbers
